Plot Extraction- Version 2
In the folder you'll see a sample data set called: sample_input3, and you also see a .csv file called Cofig3 for this data set.
Please open this file as this is how your configuration file should exactly look like.

In Cofig3:
You'll see I assigned a section for plot size,  One for the gaps/distances between rows/column

Note, when you have several section in your field, hence there is a bigger gap between those columns, you can enter the number of those columns on line 9, and the gap measurements on line 8.

Then the last session asks for the coordinate of the first/last plot. First plot is the one who has the smallest number in row/column, and last plot is the one with the biggest numbers.

This program works based on row/column information and the coordinates. Unlike the previous version that I had that required to have a modified field map, this one is very simple.

 On line 13, enter the name of your input data and run this code in windows power-shell (Or commend line) by entering this line:
python csv2kml.py Config3.csv

***Some note to use Python/powershell(or cmd)

-Put your dataset in the same folder as the python code: (......\csv-kml_v4\csv-kml_v4\CSV_KML)
-Type powershell in your windows (or type cmd)
-Define the directory to the folder:  type cd F:\Plot_extraction_method\csv-kml-Python3\CSV_KML_v4
(Give the path that your folder is stored)
-now type: python csv2kml.py Config3.csv
The code should start running (It takes about 1,2 seconds, or less) and the output in kml format will be saved in the same folder.

 
I have also included a version with Python 3, in case that's what you have installed on your system, to make sure which one you have simply type Python in the commend line.